Ensure that blocking sizes can work out (fixes GH-23)
Every QPhiX operation has to construct a `Geometry` object at some time.
Therefore the check seems to be best placed there.

At startup, there will be an uncaught exception that causes the calling
program (Chroma, tmLQCD) to crash right away. Currently Chroma will
crash because the residuals become too high. For the user it is probably
nicer to have a hard error message that points to the cause of the
problem.